Refactor to package org.kafkaquery
