+++ README.md
@@ -1,6 +1,6 @@
-# CodeFeedr KafkaQuery
+# KafkaQuery
 
-CodeFeedr KafkaQuery allows to process JSON data stored in Kafka with the help of [Flink SQL](https://ci.apache.org/projects/flink/flink-docs-release-1.12/dev/table/sql/queries.html#operations) in a streaming fashion.
+KafkaQuery allows to process JSON data stored in Kafka with the help of [Flink SQL](https://ci.apache.org/projects/flink/flink-docs-release-1.12/dev/table/sql/queries.html#operations) in a streaming fashion.
 
 
 [![asciinema page](docs/UsageExamples/Showcase.gif)](https://asciinema.org/a/388424)
