--- src/main/scala/org/codefeedr/kafkaquery/CLI.scala
@@ -1,15 +1,0 @@
-package org.codefeedr.kafkaquery
-
-import org.codefeedr.kafkaquery.parsers.Parser
-
-object CLI {
-
-  /** Main method which looks up which arguments to parse.
-    *
-    * @param args command line arguments
-    */
-  def main(args: Array[String]): Unit = {
-
-    new Parser().parse(args)
-  }
-}
